requirements.txt
pytubemp3
tkinter
customtkinter
packaging